feat!: changed prices/amounts to numeric and configured pg numeric values parsing; controller tests

+++ migrations/20210330142014_create-packages.js
@@ -3,8 +3,8 @@ exports.up = function(knex) {
     CREATE TABLE IF NOT EXISTS packages (
       id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
       name TEXT UNIQUE NOT NULL,
-      price DECIMAL(12, 2) NOT NULL DEFAULT 0,
-      credits INTEGER NOT NULL DEFAULT 0
+      price DECIMAL(8, 2) NOT NULL DEFAULT 0,
+      credits DECIMAL(12, 2) NOT NULL DEFAULT 0
     );
   `)
 }
