+++ migrations/20210330142020_create-subscriptions.js
@@ -4,7 +4,7 @@ exports.up = function(knex) {
       id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
       name TEXT UNIQUE NOT NULL,
       price DECIMAL(12, 2) NOT NULL DEFAULT 0,
-      credits INTEGER NOT NULL DEFAULT 0,
+      credits DECIMAL(12, 2) NOT NULL DEFAULT 0,
       periodicity TEXT NOT NULL DEFAULT '1 month'
     );
   `)
